requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.9.1+cu128
torchvision==0.24.1+cu128
torchaudio==2.9.1+cu128
scikit-learn>=1.3.0
albumentations>=1.3.0
tqdm>=4.66.0
pandas>=2.1.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
timm>=0.9.12
